+++ client/components/DrawNode.js
